refactor(scrcpy): separate parts depend on ADB
