+++ .vscode/settings.json
@@ -28,6 +28,7 @@
         "Nalu",
         "opendir",
         "PKCS",
+        "ponyfill",
         "runtimes",
         "Scrcpy",
         "sendrecv",
