+++ common/config/rush/repo-state.json
@@ -1,4 +1,4 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-  "pnpmShrinkwrapHash": "084447760927a169dd324c486492ee2effdc3820"
+  "pnpmShrinkwrapHash": "3798fc5f6f6c673b1888e77923d3eeace5651923"
 }
