+++ libraries/scrcpy/package.json
@@ -45,17 +45,28 @@
         "@jest/globals": "^28.1.0",
         "@types/dom-webcodecs": "^0.1.3",
         "@yume-chan/ts-package-builder": "^1.0.0",
-        "gh-release-fetch": "^2.0.4",
         "jest": "^28.1.0",
-        "tinyh264": "^0.0.7",
-        "typescript": "4.7.2",
-        "yuv-buffer": "^1.0.0",
-        "yuv-canvas": "^1.2.7"
+        "typescript": "4.7.2"
     },
     "peerDependencies": {
         "@types/dom-webcodecs": "^0.1.3",
-        "tinyh264": "^0.0.6",
+        "gh-release-fetch": "^3.0.2",
+        "tinyh264": "^0.0.7",
         "yuv-buffer": "^1.0.0",
-        "yuv-canvas": "^1.2.7"
+        "yuv-canvas": "^1.2.11"
+    },
+    "peerDependenciesMeta": {
+        "gh-release-fetch": {
+            "optional": true
+        },
+        "tinyh264": {
+            "optional": true
+        },
+        "yuv-buffer": {
+            "optional": true
+        },
+        "yuv-canvas": {
+            "optional": true
+        }
     }
 }
