+++ libraries/scrcpy/tsconfig.json
@@ -17,7 +17,7 @@
             "path": "../event/tsconfig.json"
         },
         {
-            "path": "../struct/tsconfig.json"
+            "path": "../struct/tsconfig.build.json"
         }
     ]
 }
